requirements.txt
llm
llm-sentence-transformers
gensim
slugify
progressist
sklearn
minicli
python-dotenv
